# [SPARK-14656][Core] Fix Benchmark.getPorcessorName() always return "Unknown processor" on Linux