@hvanhovell 2016-04-19
LGTM - This works nicely on Ubuntu 15.10

I'll merge this in a bit, so others have the opportunity to comment.