@hvanhovell 2016-04-19
Merging to master. Thanks!